Revert "Revert "Added b0t.functions.php ""
